--- src/entities/controllers/file-system-controller-functions/load.ts
@@ -1,77 +1,0 @@
-import { get } from "svelte/store";
-import { InjectedWindow } from "../../../common/types/injected-window";
-import { Project } from "../../models/project";
-import { SystemConfiguration } from "../../models/system-configuration";
-import projectsStore from "../../stores/projects-store";
-import { systemConfigurationsStore, SystemConfigurationStore } from "../../stores/system-configurations-store";
-
-export class FSReadFunctions {
-  private static fileApi = (window as InjectedWindow).fileApi;
-
-  public static getProjectsList = this.fileApi.getAvailableProjects;
-
-  private static async importProject (projectName : string) : Promise<Project> {
-    const projectInfo = await this.fileApi.getProjectInfo(projectName);
-    if(!projectInfo) throw Error("No config file found for project: " + projectName);
-    return new Project(projectInfo);
-  }
-
-  private static async importVersion (parentProject : Project, version : string) : Promise<SystemConfiguration> {
-    const versionInfo = await this.fileApi.getVersion(parentProject.projectName, version);
-    return new SystemConfiguration(versionInfo, parentProject.identifier);
-  }
-
-  /** Loads project from saved files */
-  public static async project (projectName : string) : Promise<Project> {
-    const project = await this.importProject(projectName);
-    projectsStore.items.update(items => {
-      const index = items.findIndex(item => item.projectName === projectName);
-      if(index !== -1) items[index] = project;
-      else items.push(project);
-      return items;
-    });
-    return project;
-  }
-
-  /** Loads project configuration (version) from saved files */
-  public static async configuration (parentProject : Project,  versionString : string) : Promise<void> {
-    const version = await this.importVersion(parentProject, versionString);
-    systemConfigurationsStore.items.update(items => {
-      const index = items.findIndex(item =>
-        get(item.version) === versionString && get(item.projectId) === parentProject.identifier);
-      if(index !== -1) items[index] = new SystemConfigurationStore(version);
-      else items.push(new SystemConfigurationStore(version));
-      return items;
-    });
-  }
-
-  public static async projectConfigurations (project : Project) : Promise<void> {
-    const versionsStrings = project.listVersions();
-    const versions : Array<SystemConfigurationStore> = [];
-    for(const version of versionsStrings) {
-      const versionInfo = await this.importVersion(project, version);
-      versions.push(new SystemConfigurationStore(versionInfo));
-    }
-    systemConfigurationsStore.items.update(items => {
-      return [
-        ...items.filter(item => get(item.projectId) !== project.identifier),
-        ...versions,
-      ];
-    });
-  }
-
-  public static async projectAndVersions (projectName : string) : Promise<void> {
-    const project = await this.project(projectName);
-    await this.projectConfigurations(project);
-  }
-
-  public static async projects () : Promise<void> {
-    const availableProjects = await this.getProjectsList();
-    for(const project of availableProjects) await this.project(project);
-  }
-
-  public static async projectsAndVersions () : Promise<void> {
-    const availableProjects = await this.getProjectsList();
-    for(const project of availableProjects) await this.projectAndVersions(project);
-  }
-}
