+++ src/entities/controllers/file-system-controller.ts
@@ -1,15 +1,7 @@
-import { FSArchiveFunctions } from "./file-system-controller-functions/archive.js";
-import { FSCreateFunctions } from "./file-system-controller-functions/update.js";
-import { FSDeleteFunctions } from "./file-system-controller-functions/delete.js";
-import { FSReadFunctions } from "./file-system-controller-functions/load.js";
+import { FSProjectFunctions } from "./file-system-controller-functions/projects.js";
+import { FSVersionsFunctions } from "./file-system-controller-functions/versions.js";
 
 export class FileSystemController {
-  public static load = FSReadFunctions;
-  public static update = FSCreateFunctions;
-  public static delete = FSDeleteFunctions;
-  public static archive = FSArchiveFunctions;
-
-
-  // Aliases Section
-  public static create = this.update;
+  public static projects = FSProjectFunctions;
+  public static versions = FSVersionsFunctions;
 }
